refactor(txpool): simplify layers and add docs
